v0.8 -  add reload function /  add auto fire /  add ammo view UI and reload UI

+++ Assets/Scripts/ScriptableObject/ProjectileDetailsSO.cs
@@ -8,7 +8,7 @@ public class ProjectileDetailsSO : ScriptableObject
     public string projectileName;
     public float projectileSpeed;
     public float projectileLifeTime;
-    public float projectileDamage;
+    public int projectileDamage;
     public GameObject projectilePrefab;
 
     [ToggleGroup("isOnEndAction")] public bool isOnEndAction;
